Calculate Simpson Effective Number of Species (Hill number q=2)

Converts Simpson's index to effective number of species using the inverse Simpson formula (1/sum(p^2)). This metric is more sensitive to dominant species than Shannon ENS.

calc_simpson_ens(abund_matrix)

Arguments

abund_matrix

Numeric matrix of abundances (rows = features, columns = samples).

Value

Numeric vector of Simpson ENS values, one per sample.

Examples

data(example_feature_table)
result <- calc_simpson_ens(as.matrix(example_feature_table[, -1]))
head(result)
#> Sample_001 Sample_002 Sample_003 Sample_004 Sample_005 Sample_006 
#>   8.353400   8.255373   8.453381   7.726897   8.232781   8.462320