Convert feature table data.frame to phyloseq object

Creates a phyloseq object from a standard feature table data.frame. Optionally incorporates taxonomy, phylogenetic tree, and sample data if provided.

to_phyloseq(
  table,
  tax_table = NULL,
  phy_tree = NULL,
  sample_data = NULL,
  feature_col = NULL
)

Arguments

table

A data.frame with feature IDs in the first column and samples as columns.

tax_table

Optional. Taxonomy table (data.frame or matrix) with feature IDs as rows and taxonomic ranks as columns (Kingdom, Phylum, Class, Order, Family, Genus, Species).

phy_tree

Optional. A phylogenetic tree (phylo object) with tip labels matching feature IDs.

sample_data

Optional. Sample metadata (data.frame) with row names matching sample names.

feature_col

Name or index of the column containing feature IDs. Default is "feature_id" or 1.

Value

A phyloseq object containing OTU table and optionally taxonomy, tree, and sample data.

Examples

if (requireNamespace("phyloseq", quietly = TRUE)) {
  data(example_feature_table)
  ps <- to_phyloseq(example_feature_table)
  print(ps)
}
#> OTU Table:          [50 taxa and 20 samples]
#>                      taxa are rows
#>         Sample_001 Sample_002 Sample_003 Sample_004 Sample_005 Sample_006
#> ASV_001       1245        892       1567        734       1123       1456
#> ASV_002        567        423        689        312        534        712
#> ASV_003        234        178        289        123        201        267
#> ASV_004         89         67        112         45         78         98
#> ASV_005       3456       2789       4123       2156       3234       3987
#> ASV_006      12345      10234      14567       8765      11234      13456
#> ASV_007       5678       4567       6789       3890       5123       6234
#> ASV_008       2345       1890       2789       1567       2123       2567
#> ASV_009        890        712       1023        567        789        934
#> ASV_010        456        367        534        278        398        478
#> ASV_011        123         98        145         76        108        132
#> ASV_012         45         34         52         27         39         48
#> ASV_013          5          3          7          2          4          6
#> ASV_014          0          1          0          0          0          1
#> ASV_015      23456      19234      26789      16789      21234      24567
#> ASV_016       8765       7234       9876       6234       7890       8901
#> ASV_017       3456       2789       4012       2456       3123       3678
#> ASV_018       1234        987       1456        789       1089       1289
#> ASV_019        567        445        656        356        489        578
#> ASV_020        234        189        267        145        201        245
#> ASV_021         89         71        102         56         78         93
#> ASV_022         34         27         39         21         29         35
#> ASV_023         12          9         14          7         10         12
#> ASV_024          3          2          4          1          2          3
#> ASV_025          1          0          1          0          0          1
#> ASV_026       4567       3789       5234       2890       4012       4789
#> ASV_027       2345       1923       2678       1456       2034       2456
#> ASV_028       1123        912       1289        689        978       1178
#> ASV_029        567        462        649        349        494        594
#> ASV_030        289        235        331        177        251        301
#> ASV_031        145        118        166         89        126        151
#> ASV_032         73         59         83         45         63         76
#> ASV_033         37         30         42         23         32         38
#> ASV_034         18         15         21         11         16         19
#> ASV_035          9          7         10          6          8         10
#> ASV_036          4          3          5          3          4          5
#> ASV_037          2          1          2          1          2          2
#> ASV_038          1          0          1          0          1          1
#> ASV_039          0          0          0          0          0          0
#> ASV_040          0          0          1          0          0          0
#> ASV_041       6789       5534       7789       4234       5901       7123
#> ASV_042       3456       2812       3967       2156       3001       3623
#> ASV_043       1789       1456       2051       1112       1551       1873
#> ASV_044        934        761       1071        580        808        978
#> ASV_045        478        389        548        296        412        498
#> ASV_046        245        200        281        152        211        255
#> ASV_047        126        103        144         78        108        131
#> ASV_048         65         53         74         40         56         67
#> ASV_049         33         27         38         21         29         34
#> ASV_050         17         14         19         11         15         17
#>         Sample_007 Sample_008 Sample_009 Sample_010 Sample_011 Sample_012
#> ASV_001        987       1234        876       1345       1098        765
#> ASV_002        456        598        387        623        498        345
#> ASV_003        156        223        134        245        189        112
#> ASV_004         56         82         48         91         71         38
#> ASV_005       2678       3456       2345       3678       3012       2234
#> ASV_006       9876      12123       9234      12789      10987       8456
#> ASV_007       4456       5567       4123       5890       5012       3789
#> ASV_008       1789       2234       1678       2456       2012       1456
#> ASV_009        645        823        598        867        723        534
#> ASV_010        323        412        289        434        356        256
#> ASV_011         89        112         82        118         97         69
#> ASV_012         32         41         29         43         36         24
#> ASV_013          3          5          2          5          4          2
#> ASV_014          0          0          0          0          0          0
#> ASV_015      18234      22345      17456      23456      20123      15678
#> ASV_016       6789       8123       6890       8567       7456       5890
#> ASV_017       2567       3234       2678       3456       2987       2234
#> ASV_018        876       1123        923       1234       1056        789
#> ASV_019        398        512        412        556        478        345
#> ASV_020        167        212        178        223        198        134
#> ASV_021         64         81         67         85         76         51
#> ASV_022         24         31         25         32         29         19
#> ASV_023          8         11          9         11         10          6
#> ASV_024          2          3          2          3          2          1
#> ASV_025          0          0          0          0          0          0
#> ASV_026       3345       4234       3456       4567       3987       3012
#> ASV_027       1678       2156       1789       2345       2012       1534
#> ASV_028        801       1034        856       1112        967        734
#> ASV_029        405        522        434        562        489        372
#> ASV_030        206        266        220        286        248        189
#> ASV_031        103        133        110        143        124         95
#> ASV_032         52         67         55         72         62         48
#> ASV_033         26         34         28         36         31         24
#> ASV_034         13         17         14         18         16         12
#> ASV_035          7          9          7          9          8          6
#> ASV_036          3          4          4          5          4          3
#> ASV_037          1          2          1          2          2          1
#> ASV_038          0          1          0          1          1          0
#> ASV_039          0          0          0          0          0          0
#> ASV_040          0          0          0          0          0          0
#> ASV_041       4890       6234       5012       6678       5789       4456
#> ASV_042       2489       3178       2551       3394       2945       2267
#> ASV_043       1284       1641       1318       1754       1521       1170
#> ASV_044        670        856        688        916        794        611
#> ASV_045        342        436        351        467        405        312
#> ASV_046        175        224        180        239        208        160
#> ASV_047         90        115         92        122        107         82
#> ASV_048         46         59         47         63         55         42
#> ASV_049         23         30         24         32         28         21
#> ASV_050         12         15         12         16         14         11
#>         Sample_013 Sample_014 Sample_015 Sample_016 Sample_017 Sample_018
#> ASV_001       1432        923       1156       1287        845       1376
#> ASV_002        678        412        545        601        378        645
#> ASV_003        256        145        198        234        128        267
#> ASV_004         95         52         73         87         42         99
#> ASV_005       3890       2567       3123       3567       2456       3789
#> ASV_006      13890       9567      11567      12890       8976      13234
#> ASV_007       6345       4234       5234       5789       4012       6123
#> ASV_008       2678       1789       2156       2345       1623       2567
#> ASV_009        945        612        756        845        578        912
#> ASV_010        489        301        378        423        267        467
#> ASV_011        138         84        105        121         73        129
#> ASV_012         49         30         37         44         26         47
#> ASV_013          6          3          4          5          2          6
#> ASV_014          1          0          0          0          0          0
#> ASV_015      25678      17890      21890      23890      16890      24890
#> ASV_016       9234       6567       7678       8234       6234       9012
#> ASV_017       3789       2567       3012       3345       2456       3678
#> ASV_018       1345        912       1067       1189        845       1312
#> ASV_019        612        401        478        534        367        589
#> ASV_020        256        156        198        223        145        245
#> ASV_021         95         60         75         84         55         92
#> ASV_022         36         22         28         32         20         35
#> ASV_023         13          8         10         11          7         13
#> ASV_024          4          2          2          3          1          3
#> ASV_025          1          0          0          0          0          1
#> ASV_026       4890       3234       4123       4456       3123       4678
#> ASV_027       2567       1678       2123       2289       1589       2456
#> ASV_028       1223        801       1012       1089        756       1178
#> ASV_029        619        405        512        549        384        594
#> ASV_030        315        206        260        279        195        301
#> ASV_031        158        103        130        140         98        151
#> ASV_032         80         52         65         70         49         76
#> ASV_033         40         26         33         35         25         38
#> ASV_034         20         13         16         18         12         19
#> ASV_035         10          7          8          9          6         10
#> ASV_036          5          3          4          4          3          5
#> ASV_037          2          1          2          2          1          2
#> ASV_038          1          0          1          1          0          1
#> ASV_039          0          0          0          0          0          0
#> ASV_040          0          0          0          0          0          0
#> ASV_041       7345       4890       6012       6456       4567       7012
#> ASV_042       3734       2489       3056       3289       2323       3567
#> ASV_043       1931       1284       1581       1701       1201       1845
#> ASV_044       1008        670        825        889        627        964
#> ASV_045        514        342        421        453        319        491
#> ASV_046        263        175        215        232        163        251
#> ASV_047        135         90        110        119         84        129
#> ASV_048         69         46         56         61         43         66
#> ASV_049         35         23         28         31         22         34
#> ASV_050         18         12         14         16         11         17
#>         Sample_019 Sample_020
#> ASV_001       1021       1198
#> ASV_002        467        523
#> ASV_003        167        212
#> ASV_004         61         79
#> ASV_005       2890       3345
#> ASV_006      10456      11890
#> ASV_007       4678       5456
#> ASV_008       1890       2234
#> ASV_009        667        801
#> ASV_010        334        398
#> ASV_011         91        114
#> ASV_012         33         40
#> ASV_013          4          5
#> ASV_014          0          0
#> ASV_015      19890      22890
#> ASV_016       7123       8012
#> ASV_017       2890       3234
#> ASV_018        978       1145
#> ASV_019        434        512
#> ASV_020        178        212
#> ASV_021         68         80
#> ASV_022         26         30
#> ASV_023          9         11
#> ASV_024          2          2
#> ASV_025          0          0
#> ASV_026       3678       4234
#> ASV_027       1890       2178
#> ASV_028        901       1045
#> ASV_029        455        527
#> ASV_030        231        268
#> ASV_031        116        134
#> ASV_032         58         67
#> ASV_033         29         34
#> ASV_034         15         17
#> ASV_035          7          8
#> ASV_036          4          4
#> ASV_037          2          2
#> ASV_038          1          1
#> ASV_039          0          0
#> ASV_040          0          0
#> ASV_041       5345       6123
#> ASV_042       2718       3112
#> ASV_043       1405       1610
#> ASV_044        734        841
#> ASV_045        374        428
#> ASV_046        192        219
#> ASV_047         99        112
#> ASV_048         51         57
#> ASV_049         26         29
#> ASV_050         13         15